Add explicit casts in comm.h and utils.h to fix narrowing warnings
Cast the size_t result in ncclTaskCollSorterInsert to uint32_t before
passing to u32fpEncode, since the value is bounded by MaxSize>>UnitLog2.
Cast chrono counts in clockRealtime to the target timespec field types
using decltype to suppress implicit narrowing from 64-bit rep types.


Mirrored-from: b492a815d4f044b63214d3583a5aeeb5a0f0da68

+++ src/include/comm.h
@@ -383,7 +383,8 @@ inline void ncclTaskCollSorterInsert(
   constexpr size_t MaxSize = ncclTaskCollSorter::MaxSize;
   constexpr int BitsPerPow2 = ncclTaskCollSorter::BitsPerPow2;
   constexpr int BinCount = ncclTaskCollSorter::BinCount;
-  int bin = u32fpEncode(std::min(MaxSize, size)>>UnitLog2, BitsPerPow2);
+  // Value is bounded by MaxSize>>UnitLog2 which fits in uint32_t
+  int bin = u32fpEncode(static_cast<uint32_t>(std::min(MaxSize, size)>>UnitLog2), BitsPerPow2);
   bin = BinCount-1 - bin; // descending bin
 
   if (me->bins[bin] == nullptr) {
